2 BYTE Permutations
The program creates the byteOutput file as well as the manchesterOutput.  Noting this is not professionally made.  Something algorithmic like this would need significant attention to optimization.  Current programs state requires a minimum of 25 seconds to run and a max of 30.  The input is just there for cosmetics at the moment.